fix: cast numpy floating/integer returns to float to satisfy mypy
np.percentile() and np.prod() return numpy scalar types (floating[Any],
signedinteger[Any]) which mypy rejects where plain float is declared.
Add explicit float() casts at the return/assignment sites in:
- var_calculator.py (calculate_historical_var, calculate_monte_carlo_var)
- advanced_risk_metrics.py (calculate_sortino_ratio)
- sentiment_analysis.py (_calc_rsi)
- risk_management_layer.py (var_95 assignment)
- stress_tester.py (all five StressTestResult constructions)

+++ quant_investor/sentiment_analysis.py
@@ -233,7 +233,7 @@ class TechnicalSentimentCalculator:
         if avg_loss < 1e-10:
             return 100.0
         rs = avg_gain / avg_loss
-        return 100.0 - 100.0 / (1.0 + rs)
+        return float(100.0 - 100.0 / (1.0 + rs))
 
     @staticmethod
     def _calc_macd(close: np.ndarray, fast: int = 12, slow: int = 26, signal: int = 9):
